requirements.txt
telethon==1.34.0
jdatetime==3.8.1
flask==2.2.5
python-dotenv